GDBus++ v4: Use glib2::Builder::AddKeyValue() feature
There are some data types wihch cannot use the glib2::Dict functions,
as GVariantDict enforces the 'a{sv}' data type.  For dictionaries
with a specific value type, such as 'a{sx}' or 'a{ss}', the
glib2::Builder API has been extended with an AddKeyValue() function
for these use cases.  Using this API, makes the C++ code much
clearer.

Signed-off-by: David Sommerseth <dazo@eurephia.org>